refactor(compose): extract media-upload logic into useMediaUploads hook
The toolbar had grown to ~455 lines mixing UI with three useHttp hooks, object-URL
lifecycle, two near-duplicate upload flows, the presigned XHR boilerplate, and the
mixing rule. Extract all of it into a useMediaUploads hook with a shared pending-chip
lifecycle (begin/fail/progress/finish), and move the raw XHR PUT into a pure
putWithProgress() helper in video.ts. The toolbar is now presentational. Also floor
(not round) client-read video duration so a 140.4s clip isn't rejected at a 140s cap.
